get latest types, refactoring
