+++ .github/workflows/build.yml
@@ -20,6 +20,8 @@ jobs:
           node-version: ${{ matrix.node-version }}
       - name: Install dependencies
         run: npm ci
+      - name: Get latest types
+        run: npm run get-latest-types
       - name: Generate types
         run: npm start
       - name: Lint types
