+++ package.json
@@ -12,6 +12,7 @@
     "compile": "tsc",
     "dtslint": "dtslint",
     "fix": "gts fix",
+    "get-latest-types": "ts-node bin/get-latest-types.ts ./types",
     "lint": "ts-node bin/lint.ts ./types",
     "prepare": "npm run compile",
     "start": "ts-node src/cli.ts --out ./types",
